Moved shooter stuff to a subsystem
